# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API APP",
        "description": "This is API for Nail App",
        "contact": {
            "email": "your-email@domain.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/notification/send-notification-store": {
            "post": {
                "tags": [
                    "notification"
                ],
                "summary": "admin send notification store",
                "operationId": "send_notification_store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "massage": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "massage": "Null",
                                    "description": "Null"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/notification/send-notification-user": {
            "post": {
                "tags": [
                    "notification"
                ],
                "summary": "admin send notification user",
                "operationId": "send_notification_user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "massage": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "massage": "Null",
                                    "description": "Null"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/notification/user": {
            "get": {
                "tags": [
                    "notification"
                ],
                "summary": "get notifications of user",
                "operationId": "get_notifications_user",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/notification/store": {
            "get": {
                "tags": [
                    "notification"
                ],
                "summary": "get notifications of store",
                "operationId": "get_notifications_store",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/notification/mark-at-read/{id}": {
            "post": {
                "tags": [
                    "notification"
                ],
                "summary": "mark at read to notification",
                "operationId": "mark_at_read",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "Bearer Token.",
                "in": "header",
                "scheme": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}